#185 - Move animation classes into Controllers/Animation namespace
Relocate SpringAxis, ZoomGeometry, IViewAnimation/IAnimationHost, and the three
IViewAnimation strategies into Display/Controllers/Animation, with the namespace
updated to FlyPhotos.Display.Controllers.Animation (folder = namespace). Add the
matching using to CanvasViewManager. Pure move; no behavior change.
